[css-pseudo] Fix position of outside 'content' markers
The position of an outside marker is calculated depending on the
'list-style-type' property. However, this patch ensures that if ::marker
has a non-normal 'content', the end of the marker will be aligned with
the start of the list item, regardless of 'list-style-type'.

The patch also simplifies the logic in InlineMarginsForOutside, since
the code for LTR and RTL directions was basically the same.

Spec: https://drafts.csswg.org/css-lists/#content-property

BUG=457718

TEST=web_tests/external/wpt/css/css-pseudo/marker-content-012.html

The test fails in legacy because 'content' doesn't work in ::marker yet.

Change-Id: I21cfd4ad3e21f2e19e8639809270bf3b4c1ad5dd