[css-pseudo] Fix WPT test mismatch expectations for ::marker
In some tests I added both "match" and "mismatch" conditions, expecting
both of them to hold since that's how Chromium treats that.

However, according to the docs, if a test has multiple conditions, they
should be treated as an OR operator. This makes a test pass in Firefox
even if the rendering is wrong. The proper way to check multiple
conditions with the AND operator is adding the extra ones in the
references.

Spec: https://web-platform-tests.org/writing-tests/reftests.html#complex-pass-conditions
Change-Id: Idb737e2486f2155f05a09a2c88bc1094c1b603e2