Have portals self-navigate with replacement.
This aligns more closely with our current thinking around session
history for portals. Maintaining a session history which is completely
disconnected from the user's real session history is of dubious value.
It's conceptually simpler to think of portals as not having a session
history (or at least having a trivial one), as this better reflects how
portals are represented to the user.

Bug: 914108
Change-Id: Ifa8f26b5dfad20a465673985d1c39ca36b2fde0f