[FlexNG] Add support for OOF-positioned static-positions within FlexNG.
This builds on previous changes to (finally!) add the correct
static-positions within the NGFlexLayoutAlgorithm.

This takes into account the main-axis, and cross-axis
alignment/justification - and converts it to the correct
inline/block axis.

As we may not know the final block-size when we encounter an
OOF-positioned node, we "adjust" this within the fragment builder once
known.

This is important for block-end/block-center aligned OOF-positioned
nodes.

Bug: 845235
Change-Id: I45c89ae1f676ec06d742b86e4b744cfc3dbe6f75