Fix the order of operations in pipeTo
Previously, Blink's implementations of pipeTo and pipeThrough did some
initialisation operations in the wrong order. This was done to simplify
the code when there were two implementations.

Now only the native implementation remains, it is simpler to do the
operations in the correct order.

In particular, switch the order of checking options w.r.t. checking the
locked status of the streams to match the standard.

Also add tests to verify the order is correct.

BUG=1031895

Change-Id: I51fbf74f4cd33ffc357a34ab302d4c1bb7b1e77b